improve vessel identity component structure
