+++ apps/fishing-map/features/reports/tabs/vessel-group-insights/VGRInsightGaps.tsx
@@ -9,11 +9,11 @@ import { VesselIdentitySourceEnum } from '@globalfishingwatch/api-types'
 import { Collapsable, Tooltip } from '@globalfishingwatch/ui-components'
 
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
+import DataTerminology from 'features/data-terminology/DataTerminology'
 import {
   GAPS_EVENTS_SOURCE_ID,
   VESSEL_GROUP_DATAVIEW_PREFIX,
 } from 'features/dataviews/dataviews.utils'
-import DataTerminology from 'features/vessel/identity/DataTerminology'
 import InsightError from 'features/vessel/insights/InsightErrorMessage'
 import VesselLink from 'features/vessel/VesselLink'
 import { formatInfoField } from 'utils/info'
