+++ apps/fishing-map/features/reports/tabs/vessel-group-insights/VGRInsights.tsx
@@ -5,11 +5,11 @@ import { DateTime } from 'luxon'
 import { Icon } from '@globalfishingwatch/ui-components'
 
 import { selectTimeRange } from 'features/app/selectors/app.timebar.selectors'
+import DataTerminology from 'features/data-terminology/DataTerminology'
 import { selectVesselsDatasets } from 'features/datasets/datasets.selectors'
 import { getDatasetLabel } from 'features/datasets/datasets.utils'
 import { formatI18nDate } from 'features/i18n/i18nDate'
 import { selectVGRVesselDatasetsWithoutEventsRelated } from 'features/reports/shared/vessels/report-vessels.selectors'
-import DataTerminology from 'features/vessel/identity/DataTerminology'
 import { MIN_INSIGHTS_YEAR } from 'features/vessel/insights/insights.config'
 
 import VesselGroupReportInsightFishing from './VGRInsightFishing'
