+++ apps/fishing-map/features/vessel/identity/VesselIdentity.module.css
@@ -127,16 +127,8 @@
 }
 
 .externalToolLinks a:not(:hover) {
-  /* filter: brightness(0) saturate(100%) invert(19%) sepia(25%) saturate(4893%) hue-rotate(206deg)
-    brightness(94%) contrast(98%); */
   opacity: var(--opacity-secondary);
   mix-blend-mode: multiply;
-}
-
-.externalToolLinks a img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
 }
 
 .disclaimer {
