+++ apps/fishing-map/features/vessel/identity/VesselIdentity.tsx
@@ -1,84 +1,21 @@
-import { useMemo } from 'react'
-import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
-import cx from 'classnames'
-import filesaver from 'file-saver'
 
-import type { RegistryExtraFieldValue, VesselRegistryOwner } from '@globalfishingwatch/api-types'
-import {
-  API_LOGIN_REQUIRED,
-  SelfReportedSource,
-  VesselIdentitySourceEnum,
-} from '@globalfishingwatch/api-types'
-import { DATASET_PRIVATE_PREFIX, VMS_DATASET_ID } from '@globalfishingwatch/datasets-client'
+import type { VesselIdentitySourceEnum } from '@globalfishingwatch/api-types'
 import type { TabsProps } from '@globalfishingwatch/ui-components'
-import { Icon, IconButton, Tabs, Tooltip } from '@globalfishingwatch/ui-components'
+import { Tabs } from '@globalfishingwatch/ui-components'
 
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
-import { isPrivateDataset } from 'features/datasets/datasets.utils'
-import { formatI18nDate } from 'features/i18n/i18nDate'
-import type { VesselLastIdentity } from 'features/search/search.slice'
-import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
-import { selectHasTMTPermission } from 'features/user/selectors/user.permissions.selectors'
-import { selectIsGFWUser, selectIsJACUser } from 'features/user/selectors/user.selectors'
-import UserLoggedIconButton from 'features/user/UserLoggedIconButton'
-import DataTerminology from 'features/vessel/identity/DataTerminology'
 import { useVesselIdentityTabs } from 'features/vessel/identity/vessel-identity.hooks'
-import VesselIdentityField from 'features/vessel/identity/VesselIdentityField'
-import VesselIdentitySelector from 'features/vessel/identity/VesselIdentitySelector'
-import { selectVesselInfoData } from 'features/vessel/selectors/vessel.selectors'
-import {
-  CUSTOM_VMS_IDENTITY_FIELD_GROUPS,
-  IDENTITY_FIELD_GROUPS,
-  REGISTRY_FIELD_GROUPS,
-  REGISTRY_SOURCES,
-  SELF_REPORTED_SECTION_SPLIT,
-} from 'features/vessel/vessel.config'
-import {
-  selectVesselDatasetId,
-  selectVesselIdentityId,
-  selectVesselIdentitySource,
-} from 'features/vessel/vessel.config.selectors'
-import { parseVesselToCSV } from 'features/vessel/vessel.download'
-import {
-  filterRegistryInfoByDateAndSSVID,
-  getCurrentIdentityVessel,
-  getLatestIdentityPrioritised,
-  getSkylightLink,
-} from 'features/vessel/vessel.utils'
-import VesselInfoCorrection from 'features/workspace/vessels/VesselInfoCorrection'
+import VesselIdentityTab from 'features/vessel/identity/VesselIdentityTab'
+import { selectVesselIdentitySource } from 'features/vessel/vessel.config.selectors'
 import { useReplaceQueryParams } from 'router/routes.hook'
-import { selectIsVesselLocation } from 'router/routes.selectors'
-import {
-  EMPTY_FIELD_PLACEHOLDER,
-  formatInfoField,
-  getVesselGearTypeLabel,
-  getVesselShipTypeLabel,
-} from 'utils/info'
-
-import VesselRegistryField from './VesselRegistryField'
-import VesselTypesField from './VesselTypesField'
 
 import styles from './VesselIdentity.module.css'
 
 const VesselIdentity = () => {
-  const { t, i18n } = useTranslation()
   const { replaceQueryParams } = useReplaceQueryParams()
-  const vesselData = useSelector(selectVesselInfoData)
-  const identityId = useSelector(selectVesselIdentityId)
-  const vesselDatasetId = useSelector(selectVesselDatasetId)
   const identitySource = useSelector(selectVesselIdentitySource)
-  const isStandaloneVesselLocation = useSelector(selectIsVesselLocation)
-  const { setTimerange } = useTimerangeConnect()
   const { identityTabs } = useVesselIdentityTabs()
-  const isGFWUser = useSelector(selectIsGFWUser)
-  const isJACUser = useSelector(selectIsJACUser)
-
-  const vesselIdentity = getCurrentIdentityVessel(vesselData, {
-    identityId,
-    identitySource,
-  })
-  const latestVesselIdentity = getLatestIdentityPrioritised(vesselData)
 
   const onTabClick: TabsProps<VesselIdentitySourceEnum>['onTabClick'] = (tab) => {
     replaceQueryParams({ vesselIdentitySource: tab.id })
@@ -89,361 +26,13 @@ const VesselIdentity = () => {
     })
   }
 
-  const onTimeRangeClick = () => {
-    setTimerange({
-      start: vesselIdentity.transmissionDateFrom,
-      end: vesselIdentity.transmissionDateTo,
-    })
-  }
-
-  const onDownloadClick = () => {
-    if (vesselIdentity) {
-      const timerange = {
-        start: vesselIdentity.transmissionDateFrom,
-        end: vesselIdentity.transmissionDateTo,
-      }
-      const { registryPublicAuthorizations, registryOwners, ssvid, shipname, flag } = vesselIdentity
-      const filteredVesselIdentity = {
-        ...vesselIdentity,
-        nShipname: formatInfoField(shipname, 'shipname') as string,
-        flag: t((t) => t[flag], {
-          defaultValue: flag,
-          ns: 'flags',
-        }) as string,
-        shiptypes: getVesselShipTypeLabel(vesselIdentity, { joinCharacter: ' -' }), // Can't be commas as it would break the csv format
-        geartypes: getVesselGearTypeLabel(vesselIdentity, { joinCharacter: ' -' }),
-        ...(identitySource === VesselIdentitySourceEnum.Registry
-          ? {
-              registryPublicAuthorizations:
-                registryPublicAuthorizations &&
-                filterRegistryInfoByDateAndSSVID(registryPublicAuthorizations, timerange, ssvid),
-              registryOwners:
-                registryOwners &&
-                (filterRegistryInfoByDateAndSSVID(
-                  registryOwners,
-                  timerange,
-                  ssvid
-                ) as VesselRegistryOwner[]),
-            }
-          : { registryPublicAuthorizations: undefined, registryOwners: undefined }),
-      }
-      const data = parseVesselToCSV(filteredVesselIdentity)
-      const blob = new Blob([data], { type: 'text/plain;charset=utf-8' })
-      filesaver.saveAs(blob, `${shipname}-${flag}.csv`)
-      trackEvent({
-        category: TrackCategory.VesselProfile,
-        action: 'vessel_identity_download',
-        label: identitySource,
-      })
-    }
-  }
-
-  const source = useMemo(() => vesselIdentity?.sourceCode, [vesselIdentity])
-
-  const identityFields = useMemo(() => {
-    const baseSource = source?.[0] as SelfReportedSource
-    const privateSource = `${baseSource}-${DATASET_PRIVATE_PREFIX}` as SelfReportedSource
-    const customIdentityFields = isPrivateDataset({ id: vesselDatasetId })
-      ? CUSTOM_VMS_IDENTITY_FIELD_GROUPS[privateSource] ||
-        CUSTOM_VMS_IDENTITY_FIELD_GROUPS[baseSource]
-      : CUSTOM_VMS_IDENTITY_FIELD_GROUPS[baseSource]
-    return customIdentityFields && customIdentityFields?.length
-      ? [...IDENTITY_FIELD_GROUPS[identitySource], ...customIdentityFields]
-      : IDENTITY_FIELD_GROUPS[identitySource]
-  }, [identitySource, source, vesselDatasetId])
-
-  const isChileanVMSVessel =
-    source?.includes(SelfReportedSource.Chile) || vesselIdentity?.flag === 'CHL'
-  const isVMSBrazilVessel = source?.includes(SelfReportedSource.Brazil)
-  const hasMoreInfo =
-    vesselIdentity?.hasComplianceInfo ||
-    vesselIdentity?.iuuStatus?.value?.toUpperCase() === 'CURRENT'
-  const hasTMTPermission = useSelector(selectHasTMTPermission)
-  const registrySourceData = REGISTRY_SOURCES.find((s) => s.key === vesselIdentity.registrySource)
-  const IdentityComponent = () => {
-    const isVMS = vesselIdentity?.sourceCode?.some((source) =>
-      source.toUpperCase().includes(VMS_DATASET_ID.toUpperCase())
-    )
-    const showSections = identitySource === VesselIdentitySourceEnum.SelfReported && !isVMS
-
-    const renderFieldGroups = (groups: typeof identityFields) =>
-      groups?.map((fieldGroup, index) => {
-        return (
-          <div key={index} className={cx(styles.fieldGroupContainer, styles.fieldGroup, {})}>
-            {fieldGroup.map((field) => {
-              let label = field.label || field.key
-              if (
-                !field.renderPlain &&
-                !isVMS &&
-                identitySource === VesselIdentitySourceEnum.SelfReported &&
-                (label === 'geartypes' || label === 'shiptypes')
-              ) {
-                label = 'gfw_' + label
-              }
-              const key = field.key as keyof VesselLastIdentity
-              let value =
-                isChileanVMSVessel && key === 'ssvid'
-                  ? EMPTY_FIELD_PLACEHOLDER
-                  : (vesselIdentity[key] as string)
-              if (key === 'depthM' || key === 'builtYear') {
-                const builtYear = vesselIdentity[key] as RegistryExtraFieldValue<number> | string
-                if (builtYear === API_LOGIN_REQUIRED) {
-                  value = API_LOGIN_REQUIRED
-                } else {
-                  value =
-                    // For registry data the builtYear is a RegistryExtraFieldValue<number>
-                    (builtYear as RegistryExtraFieldValue<number>)?.value?.toString() ||
-                    // but for VMS selfReported is a string 🤷‍♂️ so need to maintain both
-                    (typeof builtYear === 'string' ? builtYear : '') ||
-                    EMPTY_FIELD_PLACEHOLDER
-                }
-              }
-              const labelTranslation = t((t: any) => t.vessel[label], { defaultValue: label })
-              return (
-                <div key={field.key}>
-                  <div className={styles.labelContainer}>
-                    <label>{labelTranslation}</label>
-                    {field.terminologyKey && !field.renderPlain && (
-                      <DataTerminology
-                        title={labelTranslation}
-                        terminologyKey={
-                          isVMSBrazilVessel &&
-                          (field.terminologyKey === 'shiptype' ||
-                            field.terminologyKey === 'geartype')
-                            ? `${field.terminologyKey}BRA`
-                            : field.terminologyKey
-                        }
-                      />
-                    )}
-                  </div>
-                  {(key === 'shiptypes' || key === 'geartypes') && !field.renderPlain ? (
-                    <VesselTypesField
-                      vesselIdentity={vesselIdentity}
-                      fieldKey={key}
-                      identitySource={identitySource}
-                    />
-                  ) : (
-                    <VesselIdentityField value={formatInfoField(value, label as any) as string} />
-                  )}
-                </div>
-              )
-            })}
-          </div>
-        )
-      })
-
-    return (
-      <div className={cx(styles.container, { [styles.tightPadding]: showSections })}>
-        <div className={cx(styles.fieldGroup)}>
-          {identitySource === VesselIdentitySourceEnum.Registry && (
-            <div>
-              <div className={styles.labelContainer}>
-                <label>{t((t) => t.vessel.registrySources)}</label>
-                <DataTerminology
-                  title={t((t) => t.vessel.registrySources)}
-                  terminologyKey="registrySources"
-                />
-              </div>
-              {source ? (
-                <VesselIdentityField
-                  tooltip={source?.filter(Boolean)?.join(', ')}
-                  className={styles.help}
-                  value={`${source?.slice(0, 3).join(', ')}${source?.length > 3 ? '...' : ''}`}
-                />
-              ) : (
-                EMPTY_FIELD_PLACEHOLDER
-              )}
-            </div>
-          )}
-          <div
-            className={cx(styles.twoCells, {
-              [styles.sectionContent]: showSections,
-              [styles.fieldGroupContainer]: showSections,
-            })}
-          >
-            <label>
-              {t((t) => t.common.date, {
-                count: 2,
-              })}
-            </label>
-            <div className={styles.timerange}>
-              <VesselIdentityField
-                value={`${formatI18nDate(vesselIdentity.transmissionDateFrom)} - ${formatI18nDate(
-                  vesselIdentity.transmissionDateTo
-                )}`}
-              />
-              {isStandaloneVesselLocation && (
-                <IconButton
-                  size="small"
-                  icon="fit-to-timerange"
-                  tooltip={t((t) => t.timebar.fitOnThisDates)}
-                  className="print-hidden"
-                  onClick={onTimeRangeClick}
-                />
-              )}
-            </div>
-          </div>
-          <div className={styles.actionsContainer}>
-            {(isJACUser || isGFWUser) && !source?.[0].includes('VMS') && <VesselInfoCorrection />}
-
-            <UserLoggedIconButton
-              type="border"
-              icon="download"
-              size="medium"
-              className="print-hidden"
-              onClick={onDownloadClick}
-              tooltip={t((t) => t.download.identityDownload)}
-              loginTooltip={t((t) => t.download.dataDownloadLogin)}
-              tooltipPlacement="top"
-            />
-          </div>
-        </div>
-        {vesselIdentity && (
-          <div className={styles.fields}>
-            {showSections ? (
-              <>
-                <div className={styles.identitySection}>
-                  <div className={styles.sectionHeader}>
-                    <label>{t((t: any) => t.vessel.selfReportedByVessel)}</label>
-                    <DataTerminology
-                      title={t((t: any) => t.vessel.selfReportedByVessel)}
-                      terminologyKey="selfReported"
-                    />
-                  </div>
-                  <div className={styles.sectionContent}>
-                    {renderFieldGroups(identityFields.slice(0, SELF_REPORTED_SECTION_SPLIT))}
-                  </div>
-                </div>
-                <div className={styles.identitySection}>
-                  <div className={styles.sectionHeader}>
-                    <label>{t((t: any) => t.vessel.gfwPredictions)}</label>
-                    <DataTerminology
-                      title={t((t: any) => t.vessel.gfwPredictions)}
-                      terminologyKey="shiptype"
-                    />
-                  </div>
-                  <div className={styles.sectionContent}>
-                    {renderFieldGroups(identityFields.slice(SELF_REPORTED_SECTION_SPLIT))}
-                  </div>
-                </div>
-              </>
-            ) : (
-              <>
-                {renderFieldGroups(identityFields)}
-                {identitySource === VesselIdentitySourceEnum.Registry &&
-                  REGISTRY_FIELD_GROUPS.map((registryField) => {
-                    return (
-                      <VesselRegistryField
-                        key={registryField.key}
-                        registryField={registryField}
-                        vesselIdentity={vesselIdentity}
-                      />
-                    )
-                  })}
-                {identitySource === VesselIdentitySourceEnum.Registry &&
-                  hasMoreInfo &&
-                  hasTMTPermission &&
-                  registrySourceData && (
-                    <div className={cx(styles.extraInfoContainer, 'print-hidden')}>
-                      <img
-                        src={registrySourceData?.logo}
-                        alt={registrySourceData?.key}
-                        className={styles.registrySourceLogo}
-                      />
-                      <Tooltip content={t((t) => t.vessel.extraInfoTooltip)}>
-                        <div>
-                          <label>{t((t) => t.vessel.extraInfo)}</label>
-                          <a href={`mailto:${registrySourceData?.contact}`} target="_blank">
-                            {registrySourceData?.contact}
-                          </a>
-                        </div>
-                      </Tooltip>
-                    </div>
-                  )}
-              </>
-            )}
-            {vesselIdentity?.ssvid && (
-              <div
-                className={cx(
-                  { [styles.identitySection]: showSections },
-                  { [styles.sectionContent]: showSections },
-                  { [styles.fieldGroupContainer]: showSections },
-                  'print-hidden'
-                )}
-              >
-                <label>{t((t) => t.common.viewIn)}</label>
-                <div className={styles.externalToolLinks}>
-                  <a
-                    href={`https://www.marinetraffic.com/${i18n.language}/data/?asset_type=vessels&mmsi=${vesselIdentity?.ssvid}`}
-                    target="_blank"
-                    onClick={() => {
-                      trackEvent({
-                        category: TrackCategory.VesselProfile,
-                        action: 'click_marine_traffic_link',
-                      })
-                    }}
-                  >
-                    Marine Traffic
-                    <Icon icon="external-link" type="default" />
-                  </a>
-                  <a
-                    href={getSkylightLink({ skylightId: latestVesselIdentity?.ssvid })}
-                    target="_blank"
-                    onClick={() => {
-                      trackEvent({
-                        category: TrackCategory.VesselProfile,
-                        action: 'click_skylight_link',
-                      })
-                    }}
-                  >
-                    Skylight
-                    <Icon icon="external-link" type="default" />
-                  </a>
-                  <a
-                    href={`https://app.triton.fish/search?name=${vesselIdentity?.imo || vesselIdentity?.ssvid || vesselIdentity?.callsign || vesselIdentity?.shipname}`}
-                    target="_blank"
-                    onClick={() => {
-                      trackEvent({
-                        category: TrackCategory.VesselProfile,
-                        action: 'click_triton_link',
-                      })
-                    }}
-                  >
-                    Triton
-                    <Icon icon="external-link" type="default" />
-                  </a>
-                  <a
-                    href={`https://cravt.imcsnet.org/browse-vessels?keywords=${vesselIdentity?.imo || vesselIdentity?.callsign || vesselIdentity?.shipname || vesselIdentity?.nShipname}`}
-                    target="_blank"
-                    onClick={() => {
-                      trackEvent({
-                        category: TrackCategory.VesselProfile,
-                        action: 'click_cravt_link',
-                      })
-                    }}
-                  >
-                    CRAVT
-                    <Icon icon="external-link" type="default" />
-                  </a>
-                </div>
-              </div>
-            )}
-          </div>
-        )}
-        <VesselIdentitySelector />
-      </div>
-    )
-  }
   return (
     <div className={styles.identityContainer}>
       <Tabs
-        tabs={identityTabs.map((t) => ({
-          ...t,
-          content: <IdentityComponent />,
-        }))}
         activeTab={identitySource}
         onTabClick={onTabClick}
         className={styles.tabsContainer}
+        tabs={identityTabs.map((tab) => ({ ...tab, content: <VesselIdentityTab /> }))}
       />
     </div>
   )
