+++ apps/fishing-map/features/vessel/identity/fields/VesselIdentityCombinedSourceField.tsx
@@ -11,10 +11,10 @@ import { Icon, Tooltip } from '@globalfishingwatch/ui-components'
 import type { VesselLastIdentity } from 'features/search/search.slice'
 import GFWOnly from 'features/user/GFWOnly'
 import { selectIsGFWUser, selectIsJACUser } from 'features/user/selectors/user.selectors'
-import VesselIdentityField from 'features/vessel/identity/VesselIdentityField'
+import VesselIdentityField from 'features/vessel/identity/fields/VesselIdentityField'
 import { EMPTY_FIELD_PLACEHOLDER, formatInfoField } from 'utils/info'
 
-import styles from './VesselIdentity.module.css'
+import styles from '../VesselIdentity.module.css'
 
 type VesselIdentityCombinedSourceFieldProps = {
   identity: VesselLastIdentity
