+++ apps/fishing-map/features/vessel/identity/fields/VesselIdentityField.module.css
@@ -1,5 +1,6 @@
 .value {
   position: relative;
+  isolation: isolate;
 }
 
 .value::before {
