+++ apps/fishing-map/features/vessel/insights/InsightFlagChanges.tsx
@@ -5,9 +5,9 @@ import { useSelector } from 'react-redux'
 import type { ParsedAPIError } from '@globalfishingwatch/api-client'
 import type { InsightResponse } from '@globalfishingwatch/api-types'
 
+import DataTerminology from 'features/data-terminology/DataTerminology'
 import { selectIsGuestUser } from 'features/user/selectors/user.selectors'
-import DataTerminology from 'features/vessel/identity/DataTerminology'
-import VesselIdentityFieldLogin from 'features/vessel/identity/VesselIdentityFieldLogin'
+import VesselIdentityFieldLogin from 'features/vessel/identity/fields/VesselIdentityFieldLogin'
 import InsightError from 'features/vessel/insights/InsightErrorMessage'
 import { formatInfoField } from 'utils/info'
 import { listAsSentence } from 'utils/shared'
