+++ apps/fishing-map/features/vessel/insights/InsightGaps.tsx
@@ -5,7 +5,7 @@ import type { ParsedAPIError } from '@globalfishingwatch/api-client'
 import type { InsightResponse } from '@globalfishingwatch/api-types'
 import { Tooltip } from '@globalfishingwatch/ui-components'
 
-import DataTerminology from 'features/vessel/identity/DataTerminology'
+import DataTerminology from 'features/data-terminology/DataTerminology'
 
 import InsightError from './InsightErrorMessage'
 import InsightGapsDetails from './InsightGapsDetails'
