+++ apps/fishing-map/features/vessel/insights/InsightMOUList.tsx
@@ -5,10 +5,10 @@ import { useSelector } from 'react-redux'
 import type { ParsedAPIError } from '@globalfishingwatch/api-client'
 import type { InsightResponse, InsightValueInPeriod } from '@globalfishingwatch/api-types'
 
+import DataTerminology from 'features/data-terminology/DataTerminology'
 import { formatI18nDate } from 'features/i18n/i18nDate'
 import { selectIsGuestUser } from 'features/user/selectors/user.selectors'
-import DataTerminology from 'features/vessel/identity/DataTerminology'
-import VesselIdentityFieldLogin from 'features/vessel/identity/VesselIdentityFieldLogin'
+import VesselIdentityFieldLogin from 'features/vessel/identity/fields/VesselIdentityFieldLogin'
 import InsightError from 'features/vessel/insights/InsightErrorMessage'
 import { formatInfoField, upperFirst } from 'utils/info'
